Fix #1124: the Source attribute for target TerragruntOptions needs to be updated for dependency
