+++ configstack/module_test.go
@@ -1,7 +1,6 @@
 package configstack
 
 import (
-	"fmt"
 	"os"
 	"testing"
 
@@ -769,53 +768,6 @@ func TestResolveTerraformModuleNoTerraformConfig(t *testing.T) {
 	actualModules, actualErr := ResolveTerraformModules(configPaths, mockOptions, mockHowThesePathsWereFound)
 	assert.Nil(t, actualErr, "Unexpected error: %v", actualErr)
 	assertModuleListsEqual(t, expected, actualModules)
-}
-
-func TestGetTerragruntSourceForModuleHappyPath(t *testing.T) {
-	t.Parallel()
-
-	testCases := []struct {
-		config   *config.TerragruntConfig
-		opts     *options.TerragruntOptions
-		expected string
-	}{
-		{mockConfigWithSource(""), mockOptionsWithSource(t, ""), ""},
-		{mockConfigWithSource(""), mockOptionsWithSource(t, "/source/modules"), ""},
-		{mockConfigWithSource("git::git@github.com:acme/modules.git//foo/bar"), mockOptionsWithSource(t, "/source/modules"), "/source/modules//foo/bar"},
-		{mockConfigWithSource("git::git@github.com:acme/modules.git//foo/bar?ref=v0.0.1"), mockOptionsWithSource(t, "/source/modules"), "/source/modules//foo/bar"},
-		{mockConfigWithSource("git::git@github.com:acme/emr_cluster.git?ref=feature/fix_bugs"), mockOptionsWithSource(t, "/source/modules"), "/source/modules//emr_cluster"},
-		{mockConfigWithSource("git::ssh://git@ghe.ourcorp.com/OurOrg/some-module.git"), mockOptionsWithSource(t, "/source/modules"), "/source/modules//some-module"},
-		{mockConfigWithSource("github.com/hashicorp/example"), mockOptionsWithSource(t, "/source/modules"), "/source/modules//example"},
-		{mockConfigWithSource("github.com/hashicorp/example//subdir"), mockOptionsWithSource(t, "/source/modules"), "/source/modules//subdir"},
-		{mockConfigWithSource("git@github.com:hashicorp/example.git//subdir"), mockOptionsWithSource(t, "/source/modules"), "/source/modules//subdir"},
-		{mockConfigWithSource("./some/path//to/modulename"), mockOptionsWithSource(t, "/source/modules"), "/source/modules//to/modulename"},
-	}
-
-	for _, testCase := range testCases {
-		// The following is necessary to make sure testCase's values don't
-		// get updated due to concurrency within the scope of t.Run(..) below
-		testCase := testCase
-		t.Run(fmt.Sprintf("%v-%s", *testCase.config.Terraform.Source, testCase.opts.Source), func(t *testing.T) {
-			actual, err := getTerragruntSourceForModule("mock-for-test", testCase.config, testCase.opts)
-			require.NoError(t, err)
-			assert.Equal(t, testCase.expected, actual)
-		})
-	}
-}
-
-func mockOptionsWithSource(t *testing.T, sourceUrl string) *options.TerragruntOptions {
-	opts, err := options.NewTerragruntOptionsForTest("mock-for-test.hcl")
-	if err != nil {
-		t.Fatalf("Error creating terragrunt options for test %v", err)
-	}
-	opts.Source = sourceUrl
-	return opts
-}
-
-func mockConfigWithSource(sourceUrl string) *config.TerragruntConfig {
-	cfg := config.TerragruntConfig{IsPartial: true}
-	cfg.Terraform = &config.TerraformConfig{Source: &sourceUrl}
-	return &cfg
 }
 
 func ptr(str string) *string {
